backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
mediapipe==0.10.9
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.3
imagehash==4.3.1
yt-dlp
reportlab==4.0.9
pydantic==2.5.3
aiofiles==23.2.1
python-dotenv==1.0.0
